logic splitted to src files
