+++ src/object+.js
@@ -1,6 +1,7 @@
-/* Backbone-style OO functions and helpers
+/* Object extensions: backbone-style OO functions and helpers...
  * (c) Vlad Balin & Volicon, 2015
- */
+ * ------------------------------------------------------------- */
+
 ( function( spec ){
     for( var name in spec ){
         Object[ name ] || Object.defineProperty( Object, name, {
@@ -10,11 +11,10 @@
             value: spec[ name ]
         });
     }
-})( {
+})({
+    // Object.assign polyfill from MDN.
     assign : function( target, firstSource ){
-        if( target == null ){
-            throw new TypeError( 'Cannot convert first argument to object' );
-        }
+        if( target == null ) throw new TypeError( 'Cannot convert first argument to object' );
 
         var to = Object( target );
         for( var i = 1; i < arguments.length; i++ ){
@@ -35,17 +35,18 @@
         return to;
     },
 
+    // Object.transform function, similar to _.mapObject
     transform : function( dest, source, fun, context ){
-        for( var name in source ){
+        for( var name in source )
             if( source.hasOwnProperty( name ) ){
                 var value = fun.call( context, source[ name ], name );
                 typeof value === 'undefined' || ( dest[ name ] = value );
             }
-        }
 
         return dest;
     },
 
+    // get property descriptor looking through all prototype chain
     getPropertyDescriptor : function( obj, prop ){
         for( var desc; !desc && obj; obj = Object.getPrototypeOf( obj ) ){
             desc = Object.getOwnPropertyDescriptor( obj, prop );
@@ -54,13 +55,17 @@
         return desc;
     },
 
-    _typeErrors : {
-        overrideMethodWithValue : function( Ctor, name, value ){
-            console.warn( '[Type Warning] Base class method overriden with value in Object.extend({ ' + name + ' : ' + value + ' }); Object =', Ctor.prototype );
-        }
-    },
-
+    // extend function in the fashion of Backbone, with extended features required by NestedTypes
+    // - supports native properties definitions
+    // - supports forward declarations
+    // - warn in case if base class method is overriden with value. It's popular mistake when working with Backbone.
     extend : (function(){
+        var error = {
+            overrideMethodWithValue : function( Ctor, name, value ){
+                console.warn( '[Type Warning] Base class method overriden with value in Object.extend({ ' + name + ' : ' + value + ' }); Object =', Ctor.prototype );
+            }
+        };
+
         function Class(){
             this.initialize.apply( this, arguments );
         }
@@ -100,7 +105,7 @@
                     valueIsFunction = typeof value === 'function';
 
                 if( baseIsFunction && !valueIsFunction ){
-                    Object._typeErrors.overrideMethodWithValue( this, name, prop );
+                    error.overrideMethodWithValue( this, name, prop );
                 }
             }
 
@@ -111,18 +116,17 @@
             var prop = Object.getPropertyDescriptor( this.prototype, name );
 
             if( prop && typeof prop.value === 'function' ){
-                Object._typeErrors.overrideMethodWithValue( this, name, prop );
+                error.overrideMethodWithValue( this, name, prop );
             }
 
             return spec instanceof Function ? { get : spec } : spec;
         }
 
-        function define( protoProps, staticProps, mixinProps ){
+        function define( protoProps, staticProps ){
             Object.transform( this.prototype, protoProps,  warnOnError, this );
             Object.transform( this,           staticProps, warnOnError, this );
 
-            mixinProps && Object.defineProperties( this.prototype, Object.transform( {}, mixinProps, preparePropSpec, this ) );
-            Object.defineProperties( this.prototype, Object.transform( {}, protoProps.properties, preparePropSpec, this ) );
+            protoProps && Object.defineProperties( this.prototype, Object.transform( {}, protoProps.properties, preparePropSpec, this ) );
 
             return this;
         }
@@ -131,13 +135,15 @@
             for( var i = 0; i < arguments.length; i++ ){
                 var Ctor = arguments[ i ];
 
-                Ctor.extend || ( Ctor.extend = extend );
-                Ctor.define || ( Ctor.define = define );
+                Ctor.extend = extend;
+                Ctor.define = define;
                 Ctor.prototype.initialize || ( Ctor.prototype.initialize = function(){} );
             }
         };
 
         extend.attach( Class );
+        extend.Class = Class;
+        extend.error = error;
 
         return extend;
     })()
